Pull diff image metadata logic out of reporting renderer
