+++ paparazzi-gradle-plugin/src/main/java/app/cash/paparazzi/gradle/reporting/DiffImage.kt
@@ -2,9 +2,7 @@ package app.cash.paparazzi.gradle.reporting
 
 internal data class DiffImage(
   val path: String, // TODO relative path
-  val base64EncodedImage: String,
-  val testClass: String,
-  val testMethod: String
+  val base64EncodedImage: String
 ) {
   val text: String
     get() = "Error displaying image for $path"
